requirements.txt
fastapi==0.115.12
gunicorn==23.0.0
h11==0.14.0
h5py==3.13.0
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kaggle==1.7.4.2
kagglehub==0.3.12
keras==2.14.0
matplotlib==3.10.1
matplotlib-inline==0.1.7
numpy==1.26.4
panadas==0.2
pandas==2.2.3
pandocfilters==1.5.1
pydantic==2.10.6
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-intel==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.34.0
langhchain
langchain_huggingface
langchain_core
langchain_communit
faiss-cpu
sentence-transformers